Circular linked list now works

+++ src/linkedList/CircularLinkedList.java
@@ -1,55 +1,89 @@
 package linkedList;
 
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
-import java.util.function.UnaryOperator;
 
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.Nullable;
 
 
 /**
- * LinkedList Class<br>
- * Defines a double linked list
+ * CircularLinkedList Class<br>
+ * Defines a circular double linked list
  *
  * @author     Zekrom
  *
  * @param  <E>
  *                 The type of the Linked list
  */
-public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
-	
+public class CircularLinkedList <@Nullable
+E>{
+
+
+	public static void main(final String[] args) throws Exception{
+		final CircularLinkedList<@Nullable Integer> list=new CircularLinkedList<>();
+		list.insertNext(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+		list.loopUntilNext((x, i)->{
+			System.out.println(x.getPrevious()+", "+x+", "+x.getNext());
+		}, 40);
+	}
+	@Nullable
+	private Node<E> root=null;
+
 	/*
-	* Not using Right
-	*/
-	
+	 * Not using Right
+	 */
+
+	private long size=0;
+
 	/**
 	 * Returns a new CircularLinkedList
 	 */
 	public CircularLinkedList(){
 		//Do nothing
 	}
-	
+
+
+	/**
+	 * @param value
+	 * @throws Exception
+	 */
+	@SuppressWarnings("null")
+	protected boolean check(final E value) throws Exception{
+		if(this.isEmpty()) {
+			this.root=new Node<>(value);
+			this.root.setNext(this.root);
+			this.root.setPrevious(this.root);
+			this.size++;
+			return true;
+		}
+		if(this.size==1){
+			this.insert(value);
+			return true;
+		}
+		return false;
+	}
+
 	/**
 	 * Removes the first element
 	 *
 	 * @return The removed element
 	 */
-	@Override
+	@SuppressWarnings("null")
 	public E deleteLeft(){
-		if(this.left==null)return null;
-		Node<E> node=this.left;
+		if(this.root==null)return null;
+		@NonNull
+		final Node<E> node=this.root;
 		if(this.size==1){
 			this.size=0;
-			this.left=null;
+			this.root=null;
 			return node.getValue();
 		}
-		this.left=this.left.getNext();
+		this.root=this.root.getNext();
 		node.delete();
 		this.size--;
 		return node.getValue();
 	}
-
 
 	/**
 	 * Removes the index from the left
@@ -58,18 +92,18 @@ public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
 	 *                   The index to remove
 	 * @return       The removed value
 	 */
-	@Override
+	@SuppressWarnings("null")
 	public E deleteLeft(final long index){
+		if(this.root==null) return null;
 		if(this.size==1){
-			Node<E> node=this.left;
+			final Node<E> node=this.root;
 			this.size=0;
-			this.left=null;
+			this.root=null;
 			return node.getValue();
 		}
 		if(index==0) return this.deleteLeft();
 
 		final Node<E> node=this.getNodeLeft(index);
-		if(node==null)return null;
 		node.delete();
 		return node.getValue();
 	}
@@ -79,17 +113,17 @@ public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
 	 *
 	 * @return The removed element
 	 */
-	@Override
 	@SuppressWarnings("null")
 	public E deleteRight(){
-		if(this.left==null)return null;
-		Node<E> node=this.left;
+		if(this.root==null)return null;
+		@NonNull
+		final Node<E> node=this.root;
 		if(this.size==1){
 			this.size=0;
-			this.left=null;
+			this.root=null;
 			return node.getValue();
 		}
-		this.left=this.left.getPrevious();
+		this.root=this.root.getPrevious();
 		node.delete();
 		this.size--;
 		return node.getValue();
@@ -102,44 +136,54 @@ public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
 	 *                   The index to remove
 	 * @return       The removed value
 	 */
-	@Override
+	@SuppressWarnings("null")
 	public E deleteRight(final long index){
+		if(this.root==null) return null;
 		if(this.size==1){
-			Node<E> node=this.left;
+			final Node<E> node=this.root;
 			this.size=0;
-			this.left=null;
+			this.root=null;
 			return node.getValue();
 		}
 		if(index==0) return this.deleteRight();
 
 		final Node<E> node=this.getNodeRight(index);
-		if(node==null)return null;
 		node.delete();
 		return node.getValue();
 	}
-	
+
 	/**
-	 * Returns the node of the index from the right
+	 * Iterates for every element in the linked list in the order they are provided, going to the Next
+	 * elements
 	 *
-	 * @param  index
-	 *                   The index to get the node
-	 * @return       The node of the index
+	 * @param consumer
+	 *                     The function to do for each element, no change is made to the value
 	 */
-	@Override
 	@SuppressWarnings("null")
-	protected @NonNull Node<E> getNodeRight(final long index){
-		index=index%this.size;
-		if(index>this.size/2) return this.getNodeLeft(this.size-index-1);
-		@NonNull
-		Node<E> next=this.left;
-		int i=0;
-		while(i<index){
-			next=next.getPrevious();
-			i++;
+	public void forEachNext(final Consumer<E> consumer){
+		Node<E> node=this.root;
+		for(int i=0; i<this.size; i++){
+			consumer.accept(node.getValue());
+			node=node.getNext();
 		}
-		return next;
 	}
-	
+
+	/**
+	 * Iterates for every element in the linked list in the order they are provided, going to the
+	 * Previous elements
+	 *
+	 * @param consumer
+	 *                     The function to do for each element, no change is made to the value
+	 */
+	@SuppressWarnings("null")
+	public void forEachPrevious(final Consumer<E> consumer){
+		Node<E> node=this.root;
+		for(int i=0; i<this.size; i++){
+			consumer.accept(node.getValue());
+			node=node.getPrevious();
+		}
+	}
+
 	/**
 	 * Returns the node of the index from the left
 	 *
@@ -147,18 +191,16 @@ public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
 	 *                   The index to get the node
 	 * @return       The node of the index
 	 */
-	@Override
 	@SuppressWarnings("null")
-	protected @NonNull Node<E> getNodeLeft(final long index){
+	protected @NonNull
+	Node<E> getNodeLeft(long index){
+		if(
+			this.root==null
+			) throw new NullPointerException("getNodeLeft Encountered a null root");
 		index=index%this.size;
-		/*
-		* //What happens if I do this?
-		* super.getNodeLeft(index%this.size);
-		* //Does it call LinkedList.getNodeRight() or CircularLinkedList.getNodeRight() <- What I want?
-		*/
 		if(index>this.size/2) return this.getNodeRight(this.size-index-1);
 		@NonNull
-		Node<E> prev=this.left;
+		Node<E> prev=this.root;
 		int i=0;
 		while(i<index){
 			prev=prev.getNext();
@@ -166,85 +208,197 @@ public class CircularLinkedList<@Nullable E> extends LinkedList <E>{
 		}
 		return prev;
 	}
-	
-	
 
-	protected void insert(final E value){
-		this.left=new Node<>(value);
-		this.left.setNext(this.left);
-		this.left.setPrevious(this.left);
-		this.size++;
+
+	/**
+	 * Returns the node of the index from the right
+	 *
+	 * @param  index
+	 *                   The index to get the node
+	 * @return       The node of the index
+	 */
+	@SuppressWarnings("null")
+	protected @NonNull
+	Node<E> getNodeRight(long index){
+		if(
+			this.root==null
+			) throw new NullPointerException(
+				"getNodeRight Encountered a null root"
+				);
+
+		index=index%this.size;
+		if(index>this.size/2) return this.getNodeLeft(this.size-index-1);
+		@NonNull
+		Node<E> next=this.root;
+		int i=0;
+		while(i<index){
+			next=next.getPrevious();
+			i++;
+		}
+		return next;
 	}
 
 	/**
-	 * Inserts the value at the beginning
+	 * Inserts the value if there is only one value
 	 *
 	 * @param value
 	 *                  The value to insert
 	 */
-	public void insertLeft(final E value){
-		if(this.left==null){
-			this.insert(value);
-			return;
-		}
-		this.left=new Node<>(this.left, value, this.left.getNext());
+	protected void insert(final E value){
+		this.root=new Node<>(this.root, value, this.root);
+		this.size++;
+	}
+
+
+	/**
+	 * Inserts the value at the root and the existing node is sifted to the next
+	 *
+	 * @param value
+	 *                  The value to insert
+	 * @throws Exception
+	 */
+	@SuppressWarnings("null")
+	public void insertNext(final E value) throws Exception{
+		if(this.check(value))return;
+		this.root=new Node<>(this.root, value, this.root.getNext());
 		this.size++;
 	}
 
 	/**
-	 * Inserts the value after the left based index
+	 * Inserts the values at the root and the existing node is sifted to the next
+	 *
+	 * @param  values
+	 *                       The values to insert
+	 * @throws Exception
+	 */
+	@SuppressWarnings("unchecked")
+	public void insertNext(final E... values) throws Exception{
+		for(final E value : values){
+			this.insertNext(value);
+		}
+	}
+
+
+	/**
+	 * Inserts the value after the next based index
 	 *
 	 * @param  index
-	 *                   The index to insert the value (Inserts after)
+	 *                       The index to insert the value (Inserts after)
 	 * @param  value
-	 *                   The value to insert
-	 * @return       Success of insertion
+	 *                       The value to insert
+	 * @throws Exception
 	 */
 	@SuppressWarnings("unused")
-	public boolean insertLeft(final int index, final E value){
-		if(index==-1||this.left==null){
-			this.insert(value);
-			return true;
-		}
+	public void insertNext(final int index, final E value) throws Exception{
+		if(this.check(value)) return;
 		@NonNull
 		final Node<E> prev=this.getNodeLeft(index);
 		new Node<>(prev, value, prev.getNext());
 		this.size++;
-		return true;
 	}
 
+	/**
+	 * Inserts the values at the root and the existing node is sifted to the previous
+	 *
+	 * @param  values
+	 *                       The values to insert
+	 * @throws Exception
+	 */
+	@SuppressWarnings("unchecked")
+	public void insertPrevious(final E... values) throws Exception{
+		for(final E value : values){
+			this.insertPrevious(value);
+		}
+	}
 
 	/**
-	 * Inserts the value at the end
+	 * Inserts the value at the root and the existing node is sifted to the previous
 	 *
-	 * @param value
-	 *                  The value to insert
+	 * @param  value
+	 *                       The value to insert
+	 * @throws Exception
 	 */
-	public void insertRight(final E value){
-		if(this.left==null){
-			this.insert(value);
-			return;
-		}
-		this.left=new Node<>(this.left.getPrevious(), value, this.left);
+	@SuppressWarnings("null")
+	public void insertPrevious(final E value) throws Exception{
+		if(this.check(value)) return;
+		this.root=new Node<>(this.root.getPrevious(), value, this.root);
 		this.size++;
 	}
 
-	/**
-	 * Inserts the value after the right based index
-	 * @param index The index to insert the value (Inserts after)
-	 * @param value The value to insert
-	 * @return Success of insertion
+	/***
+	 * Inserts the value after the previous based index
+	 *
+	 * @param  index
+	 *                       The index to insert the value (Inserts after)
+	 * @param  value
+	 *                       The value to insert
+	 * @throws Exception
 	 */
 	@SuppressWarnings("unused")
-	public boolean insertRight(final int index, final E value){
-		if(index==-1||this.left==null){
-			this.insert(value);
-			return true;
-		}
+	public void insertPrevious(final int index, final E value) throws Exception{
+		if(this.check(value)) return;
 		@NonNull
 		final Node<E> next=this.getNodeRight(index);
 		new Node<>(next.getPrevious(), value, next);
 		this.size++;
-		return true;
 	}
+
+	/**
+	 * Tests if the linked list is empty
+	 *
+	 * @return           The state of the array
+	 * @throws Exception
+	 */
+	public boolean isEmpty() throws Exception{
+		if(this.size==0&&this.root==null) return true;
+		if(this.size==0||this.root==null) throw new Exception();
+		return false;
+	}
+
+	/**
+	 * Loops until max-1
+	 *
+	 * @param  consumer
+	 *                       The function to do on each iteration
+	 * @param  max
+	 *                       The amount of times to call the function
+	 * @throws Exception
+	 */
+	@SuppressWarnings("null")
+	public void loopUntilNext(
+		final BiConsumer<Node<E>, Long> consumer, final long max
+		) throws Exception{
+		if(this.isEmpty()) return;
+		@NonNull
+		Node<E> node=this.root;
+		for(long i=0; i<max; i++){
+			consumer.accept(node, i);
+			node=node.getNext();
+		}
+	}
+
+	/**
+	 * Returns a string representation of the object
+	 */
+	@SuppressWarnings("null")
+	@Override
+	public String toString(){
+		try{
+			if(this.isEmpty()) return "CircularLinkedList[]";
+		}
+		catch(final Exception e){
+			e.printStackTrace();
+		}
+
+		final StringBuilder builder=new StringBuilder();
+
+		Node<E> node=this.root;
+		for(int i=0; i<this.size; i++){
+			builder.append(node).append(", ");
+			node=node.getNext();
+		}
+		builder.append("...]");
+		return builder.toString();
+	}
+
 }
